Move role functionality into RolesManager.
