+++ lib/galaxy/managers/roles.py
@@ -2,16 +2,43 @@
 Manager and Serializer for Roles.
 """
 import logging
+from typing import List, Optional
 
+from pydantic import (
+    BaseModel,
+    Field,
+)
 from sqlalchemy import false
 from sqlalchemy.orm import exc as sqlalchemy_exceptions
 
 import galaxy.exceptions
 from galaxy import model
+from galaxy.exceptions import RequestParameterInvalidException
 from galaxy.managers import base
+from galaxy.schema.fields import EncodedDatabaseIdField
 from galaxy.util import unicodify
 
 log = logging.getLogger(__name__)
+
+RoleIdField = Field("ID", description="Encoded ID of the role")
+RoleNameField = Field(title="Name", description="Name of the role")
+RoleDescriptionField = Field(title="Description", description="Description of the role")
+
+
+class RoleModel(BaseModel):
+    id: EncodedDatabaseIdField = RoleIdField
+    name: str = RoleNameField
+    description: str = RoleDescriptionField
+    type: str = Field(title="Type", description="Type or category of the role")
+    url: str = Field(title="URL", description="URL for the role")
+    model_class: str = Field(title="Model class", description="Database model class (Role)")
+
+
+class RoleDefeinitionModel(BaseModel):
+    name: str = RoleNameField
+    description: str = RoleDescriptionField
+    user_ids: Optional[List[EncodedDatabaseIdField]] = Field(title="User IDs", default=[])
+    group_ids: Optional[List[EncodedDatabaseIdField]] = Field(title="Group IDs", default=[])
 
 
 class RoleManager(base.ModelManager):
@@ -60,3 +87,30 @@ class RoleManager(base.ModelManager):
             if trans.user_is_admin or trans.app.security_agent.ok_to_display(trans.user, role):
                 roles.append(role)
         return roles
+
+    def create(self, trans, role_definition_model):
+        name = role_definition_model.name
+        description = role_definition_model.description
+        user_ids = role_definition_model.user_ids or []
+        group_ids = role_definition_model.group_ids or []
+
+        if trans.sa_session.query(trans.app.model.Role).filter(trans.app.model.Role.table.c.name == name).first():
+            raise RequestParameterInvalidException(f"A role with that name already exists [{name}]")
+
+        role_type = trans.app.model.Role.types.ADMIN  # TODO: allow non-admins to create roles
+
+        role = trans.app.model.Role(name=name, description=description, type=role_type)
+        trans.sa_session.add(role)
+        users = [trans.sa_session.query(trans.model.User).get(trans.security.decode_id(i)) for i in user_ids]
+        groups = [trans.sa_session.query(trans.model.Group).get(trans.security.decode_id(i)) for i in group_ids]
+
+        # Create the UserRoleAssociations
+        for user in users:
+            trans.app.security_agent.associate_user_role(user, role)
+
+        # Create the GroupRoleAssociations
+        for group in groups:
+            trans.app.security_agent.associate_group_role(group, role)
+
+        trans.sa_session.flush()
+        return role
